Got rid of TopButtons component, kept MenuButtons component
